Fixed 32bits/64bits conversions (fixed #162)

+++ src/data-types/mailstream_cancel.c
@@ -174,7 +174,7 @@ void mailstream_cancel_notify(struct mailstream_cancel * cancel)
   
   ch = 0;
 #ifndef WIN32
-  r = write(cancel->ms_fds[1], &ch, 1);
+  r = (int) write(cancel->ms_fds[1], &ch, 1);
 #else
   SetEvent(ms_internal->event);
 #endif
@@ -185,7 +185,7 @@ void mailstream_cancel_ack(struct mailstream_cancel * cancel)
 #ifndef WIN32
   char ch;
   int r;
-  r = read(cancel->ms_fds[0], &ch, 1);
+  r = (int) read(cancel->ms_fds[0], &ch, 1);
 #endif
 }
